\name{mkdate_merge}
\alias{mkdate_merge}
\title{making year-month date date Dataset}
\usage{
mkdate_merge()
}
\description{
making year-month date Dataset
}
\examples{

# making date

# Input: Year=2023 and Month=4 // Output: DATE_merge=2023-04-01
  df<-mkdate_merge(df, column number of year, column number of month)
  df<-mkdate_merge(df, 4, 5)

}